# perf: no deep cloning