An alternative to #71 which removes the need for any deep cloning.

All unit & spec tests in the `jsonld-streaming-parser` pass with these changes.

I have also done a (small) benchmark of these changes. As can be seen the time taken to parse a context with these changes significantly decreases.

Before:
Parse a context that has not been cached; and without caching in place x 78.17 ops/sec ±0.73% (78 runs sampled)
Parse a context object that has not been cached x 2,280 ops/sec ±0.65% (91 runs sampled)

After:
Parse a context that has not been cached; and without caching in place x 123 ops/sec ±0.78% (84 runs sampled)
Parse a context object that has not been cached x 5,417 ops/sec ±0.70% (87 runs sampled)